Login

+++ app/main/views.py
@@ -7,6 +7,7 @@ from .. import db,photos
 from . import main
 from app.utils import Toppings
 from app.email import mail_message
+from flask_login import current_user, login_required
 
 
 @main.route('/')
@@ -15,8 +16,12 @@ def index():
     return render_template('index.html',all_pizzas=all_pizzas)
 
 @main.route('/add',methods = ["GET","POST"])
+@login_required
 def add():
     """Add pizza function"""
+    if not current_user.is_admin:
+        abort(403)
+
     all_pizzas=Pizza.query.all()
 
     if request.method=='POST':
@@ -40,8 +45,11 @@ def add():
 
 
 @main.route('/pizza/edit/<int:id>', methods=['GET', 'POST'])
+@login_required
 def edit(id):
     """Edit pizza function"""
+    if not current_user.is_admin:
+        abort(403)
 
     all_data=Pizza.query.get(id)
     if request.method=='POST':        
@@ -59,8 +67,12 @@ def edit(id):
 
 
 @main.route('/delete/<id>',methods=["GET","POST"])
+@login_required
 def delete(id):
     """Delete pizza function"""
+    if not current_user.is_admin:
+        abort(403)
+
     data=Pizza.query.get(id)
     db.session.delete(data)
     db.session.commit()
@@ -72,6 +84,7 @@ def delete(id):
 
 # order
 @main.route('/order/<pizza_id>', methods=["GET", "POST"])
+@login_required
 def order(pizza_id):
     '''Order pizza'''
     pizza = Pizza.query.get(pizza_id)
@@ -84,13 +97,13 @@ def order(pizza_id):
         topping_price = int(Toppings.get_topping(form.topping.data).price)
         total = (int(pizza.price) + topping_price) * int(form.quantity.data)
 
-        order = Order(total=total, pizza=pizza, quantity = form.quantity.data, topping=form.topping.data)
+        order = Order(total=total, pizza=pizza, quantity = form.quantity.data, topping=form.topping.data, user=current_user)
 
         db.session.add(order)
         db.session.commit()
 
         #send email to user
-        # mail_message("Order Successsful","email/order-success",current_user.email,user=current_user)
+        mail_message("Order Successsful","email/order-success",current_user.email,user=current_user)
 
         return redirect(url_for('.order_success', order_id = order.id))
 
@@ -98,6 +111,7 @@ def order(pizza_id):
 
 
 @main.route('/order-success/<order_id>', methods=["GET"])
+@login_required
 def order_success(order_id):
     order = Order.query.get(order_id)
     if not order:
